Reorganized folder structure with additional security and MERN features
