+++ .babelrc
@@ -1,0 +1,1 @@
+{ "presets": ["es2015", "stage-2"] }
