+++ .gitignore
@@ -1,3 +1,4 @@
 
 # Dependency directories
-node_modules
+package-lock.json
+node_modules/
