+++ controller/index.js
@@ -1,0 +1,4 @@
+import userController from './user.controller';
+export {
+   userController
+};
