--- init.jsx
@@ -1,1 +1,0 @@
-hello world
